* commit '5e898d2a869c05fe4a954bab9f643026cf5ca638':
  DM-10380 Fixed node and angularjs to use tsc and created readme for Angular
  DM-10380 Tweaked build.sh's for ts clients. Removed sdk from herd-docs and the docs for it are generated in herd-sdk.  Need to investigate generating docs from sdk and not using hte plugin ourselves.

+++ herd-code/herd-sdk/src/tsconfig/node/build.sh
@@ -1,19 +1,30 @@
-# Clean up previous distributions
-rm -rf dist
-rm -rf build
+if [ ! -d "dist" ]; then
+  echo "cleaning dist"
+  rm -rf dist
+fi
+
+if [ ! -d "build" ]; then
+  echo "cleaning build"
+  rm -rf build
+fi
 
 if [ ! -d "node_modules" ]; then
+  echo "installing node modules"
   npm install
 fi
 
-# Variable pointing to NGC
-TSC="node_modules/.bin/ngc"
+# Variable pointing to TSC
+TSC="node_modules/.bin/tsc"
 ROLLUP="node_modules/.bin/rollup"
 
+echo "Running TSC on tsconfig-build.json"
 $TSC -p tsconfig-build.json
+echo "Running creating ROLLUP bundle dist/api.js"
 $ROLLUP build/api.js --output.file dist/api.js --output.format es
 
+echo "Running TSC on tsconfig-build.es5.json"
 $TSC -p tsconfig-build.es5.json
+echo "Running creating ROLLUP bundle dist/api.es5.js"
 $ROLLUP build/api.js --output.file dist/api.es5.js --output.format es
 
 rsync -a --exclude=*.js build/ dist
